Use base search for known DN lookup
Performing a 'base' search using the DN as base is the standard way.

Using (distinguishedName=) as a filter is not universal – for example,
OpenLDAP uses entryDN instead.

And while OpenLDAP 2.6.x seems to *recognize* distinguishedName, it
gives a completely different result (applying the match to all
DN-containing attributes rather than to the entry's own DN) – so the
search gives 25 entries, none of which were the correct one.