Add support for SASL bind
We introduce a ldap_sasl() method in Ldap and LdapMulti with
an optionnal array of SASL options. A module that subclasses
Ldap or Ldapmulti can use it instead of simple login().

This requires SASL bind support in Symfony, which has been
merged in the 7.3 branch. SimpleSAMLphp uses Symfony 7.2.
How should this ne handled? I can backport the patches
for Symfony 7.2, but do we have a way to fold them in
the simpleSAMLphp package?